# Separate metadata and incomplete job tracking