Very large numbers of metadata files in the .snakemake/metadata folder can lead to slow performance on network file systems (e.g. Lustre) when checking if jobs are incomplete (--rerun-incomplete). 

1) This code separates the metadata and job-incomplete tracking, by creating an extra .snakemake/incomplete folder.
For the latter, it implements caching, as a simple file existence check suffices to determine if a job is incomplete.  The number of tracking files in the incomplete folder usually remains low, as these are deleted after a job finishes. This results in large speedups when combining --rerun-incomplete with a workflow with many files on a network file system.

Code has been added to enable backwards compatibility with the current .snakemake folder structure.

2) The PR adds also a command line option to make storing of persistent metadata optional to reduce the small file load.