@mhulsman 2020-10-16
Also did a quick benchmark of this code, using the same 283k workflow as mentioned in the other pull request, but then not in an empty dir, but one at a point when it is almost complete. Total files will probably be >500k. 

Old: 30985 seconds,   of which 20311 seconds was spent in '_read_record_uncached'. 
New: 776 seconds, of which 21 seconds was spent 'check_incomplete' and 'cache_incomplete_folder'. 

Remaining difference is mostly due to IOCache framework: 
old: 8361 seconds in io.py:wrapper (of which 8271 in 'is_newer')
new:  69 seconds in io.py:wrapper.