@hwalinga 2024-10-02
I think this PR caused #2933

The IncompleteFilesException suggest removing metadata but --cleanup-metadata does not seem to go over the new incomplete folder as well.